requirements.txt
# Core parsing library
lark-parser==1.1.7

# XML generation and manipulation
lxml==4.9.3

# CLI framework
click==8.1.7

# Testing framework
pytest==7.4.2
pytest-cov==4.1.0

# Code formatting and linting
black==23.9.1
flake8==6.1.0

# Type checking
mypy==1.6.1

# Development utilities
pre-commit==3.4.0
